git subrepo pull --branch=master --force src/Eng
subrepo:
  subdir:   "src/Eng"
  merged:   "813a2520"
upstream:
  origin:   "Eng"
  branch:   "master"
  commit:   "813a2520"
git-subrepo:
  version:  "0.4.6"
  origin:   "???"
  commit:   "???"

+++ src/Eng/first-party/Net/VarContainer.h
@@ -17,11 +17,11 @@ namespace Net {
         template<class T>
         void SaveVar(const Var<T> &v) {
             assert(CheckHashes(v.hash_.hash));
-            auto var_beg = (int_type) data_bytes_.size();
+            auto var_beg = int_type(uint16_t(data_bytes_.size()));
             header_.push_back(v.hash_.hash);
             header_.push_back(var_beg);
 
-            auto data_beg = (int_type) data_bytes_.size();
+            auto data_beg = int_type(uint16_t(data_bytes_.size()));
             data_bytes_.resize(var_beg + sizeof(T));
 
             memcpy(&data_bytes_[data_beg], v.p_val(), sizeof(T));
